Politica de acceso parte 1. Se elimina el acceso al borrado para las publicaciones que no sean del mismo usuario que las creo.

+++ publicaciones/publicaciones/app/Http/Controllers/PostController.php
@@ -31,7 +31,11 @@ class PostController extends Controller
        return back()->with('status','publicacion guardada correctamente');
     }
 
-    public function destroy(Post $post){
+    public function destroy(Request $request, Post $post){
+        //dd($request->user()->id);
+        if($request->user()->id !== $post->user_id){
+            abort(403);
+        }
         $post->delete();
         return back();
      }
